% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counties.R
\docType{data}
\name{counties}
\alias{counties}
\title{Shapes of 26 counties in the Republic of Ireland.}
\format{A data frame with 57,961 rows and 7 variables:
\describe{
  \item{long}{Longitude}
  \item{lat}{Latitude}
  \item{order}{Vertex order}
  \item{hole}{Vertices assoicated with holes in county shapes}
  \item{piece}{Numbers for different pieces of individual counties}
  \item{County}{County Name. Can be matched to Census Data and certain other datasets.}
}}
\source{
\url{https://data-osi.opendata.arcgis.com/}
}
\usage{
counties
}
\description{
A database containing shapes of 26 counties in the
Republic of Ireland. To be used for producing maps
of Census data.
}
\details{
This data is dervied from shapefiles created by Ordnance
Survey Ireland (OSi), and taken from their Open Data
Portal. The OSi has released this data under a
standard Creative Commons Licence. For more information
see the README file associated with this package.
}
\keyword{datasets}
