% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_titles.R
\docType{data}
\name{keep_titles}
\alias{keep_titles}
\title{Database of CSO tables.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3295 rows and 2 columns.}
\usage{
keep_titles
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\description{
Database of CSO tables.
}
\examples{
add(1, 1)
add(10, 1)
}
\keyword{datasets}
