% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cso_database.R
\docType{data}
\name{cso_database}
\alias{cso_database}
\title{Database of all CSO data tables.}
\format{A data frame with many rows and 2 variables:
\describe{
  \item{Item}{Full name of data table}
  \item{Code}{The 5 character code for this table}
}}
\source{
\url{https://www.cso.ie/webserviceclient/DatasetListing.aspx}
}
\usage{
cso_database
}
\description{
A database containing all of the data tables included
on the CSO Statbank website.
}
\keyword{datasets}
